+++ simple-mdns/tests/service_discovery_tests.rs
@@ -29,13 +29,13 @@ fn service_discovery_can_find_services() {
 
     std::thread::sleep(Duration::from_secs(1));
 
-    let mut service_discovery_a = ServiceDiscovery::new("_srv3._tcp.com", 360).unwrap();
-    let mut service_discovery_b = ServiceDiscovery::new("_srv3._tcp.com", 360).unwrap();
+    let mut service_discovery_a = ServiceDiscovery::new("_srv3._tcp.com", 10).unwrap();
+    let mut service_discovery_b = ServiceDiscovery::new("_srv3._tcp.com", 10).unwrap();
 
     service_discovery_a.add_socket_address("192.168.1.2:8080".parse().unwrap());
     service_discovery_b.add_socket_address("192.168.1.3:8080".parse().unwrap());
 
-    std::thread::sleep(Duration::from_secs(3));
+    std::thread::sleep(Duration::from_secs(15));
 
     let mut services_a = service_discovery_a.get_known_services();
     let mut services_b = service_discovery_b.get_known_services();