+++ simple-mdns/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Simple MDNS
 
+## 0.2.1
+
+Fix a bug that caused the refresh on the service discovery to panic
+Increase socket2 version number
+
 ## 0.2.0
 
 Remove Futures and Tokio dependencies by switching to a simple thread implementation